DD-2889 Rename 'yEarn' to 'yearn'
