+++ config/arbitrum/stable-testnet.json
@@ -1,9 +1,9 @@
 {
   "subgraphName": "in19farkt/delphi-stable-staging-arbitrum",
-  "startBlock": 13466938,
+  "startBlock": 2465385,
   "network": "arbitrum-rinkeby",
   "modules": {
     "basisVaultsRegistry": "0x0000000000000000000000000000000000000000"
   },
-  "callSupported": true
+  "callSupported": false
 }
