+++ config/bsc/stable-mainnet.json
@@ -1,6 +1,6 @@
 {
   "subgraphName": "in19farkt/delphi-stable-mainnet-bsc",
-  "startBlock": 13466938,
+  "startBlock": 12015134,
   "network": "bsc",
   "modules": {
     "basisVaultsRegistry": "0x0000000000000000000000000000000000000000"
