+++ config/eth/stable-testnet.json
@@ -3,7 +3,7 @@
   "startBlock": 7004100,
   "network": "rinkeby",
   "modules": {
-    "yEarnVaultSavingsV2": "0x90AA1e10C11B3B97CC82C17478acbf51DaDFbb3d",
+    "yearnVaultSavingsV2": "0x90AA1e10C11B3B97CC82C17478acbf51DaDFbb3d",
     "akroStakingPool": "0x6887DF2f4296e8B772cb19479472A16E836dB9e0",
     "adelStakingPool": "0x1a80540930B869df525981c4231Cd06b0Dcbf668"
   },
