+++ eth-subgraph.template.yaml
@@ -20,18 +20,18 @@ dataSources:
           file: ./abis/ERC20Detailed.json
         - name: StakingPool
           file: ./abis/StakingPool.json
-        - name: YEarnVaultSavings
-          file: ./abis/YEarnVaultSavings.json
+        - name: YearnVaultSavings
+          file: ./abis/YearnVaultSavings.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
       file: ./src/mappings/shared.ts
   - kind: ethereum/contract
-    name: YEarnVaultSavings
+    name: YearnVaultSavings
     network: {{network}}
     source:
-      address: "{{modules.yEarnVaultSavingsV2}}"
-      abi: YEarnVaultSavings
+      address: "{{modules.yearnVaultSavingsV2}}"
+      abi: YearnVaultSavings
       startBlock: {{startBlock}}
     mapping:
       kind: ethereum/events
@@ -40,8 +40,8 @@ dataSources:
       entities:
         - User
       abis:
-        - name: YEarnVaultSavings
-          file: ./abis/YEarnVaultSavings.json
+        - name: YearnVaultSavings
+          file: ./abis/YearnVaultSavings.json
         - name: ERC20Detailed
           file: ./abis/ERC20Detailed.json
       eventHandlers:
@@ -53,7 +53,7 @@ dataSources:
           handler: handleVaultActivated
         - event: Deposit(indexed address,indexed address,uint256,uint256)
           handler: handleDeposit
-      file: ./src/mappings/yEarnVaultSavingsV2.ts
+      file: ./src/mappings/yearnVaultSavingsV2.ts
   - kind: ethereum/contract
     name: AKROStakingPool
     network: {{network}}
@@ -104,7 +104,7 @@ dataSources:
       file: ./src/mappings/adelStakingPool.ts
 templates:
   - kind: ethereum/contract
-    name: YEarnVaultV2
+    name: YearnVaultV2
     network: {{network}}
     source:
       abi: ERC20Detailed
@@ -120,4 +120,4 @@ templates:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ./src/mappings/yEarnVaultV2.ts
+      file: ./src/mappings/yearnVaultV2.ts
