+++ src/entities/globalStats/deactivateUserIfZeroBalance.ts
@@ -4,7 +4,7 @@ import { User } from "../../../generated/schema";
 import { loadGlobalStat } from "./loadGlobalStats";
 
 export function deactivateUserIfZeroBalance(user: User): void {
-  let areAllUserPoolsEmpty = !user.stakingPools.length && !user.yEarnVaults.length && !user.basisVaults.length;
+  let areAllUserPoolsEmpty = !user.stakingPools.length && !user.yearnVaults.length && !user.basisVaults.length;
   if (user.active && areAllUserPoolsEmpty) {
     let globalStat = loadGlobalStat();
     globalStat.activeMembersCount -= 1;
